+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * First release!
 * Added favicon.
 * Added Bootstrap styles to pages that didn't have them yet.
+* Added a warning to the advanced search box to note that it's not yet 100% useful.
 
 ## 0.2.1